1. shader parsing has been added. 2. extract from DDS to TGA and save near exported FBX added 3. material and textures linkage in exported FBX added
Attention - new 3rdparty library has been added. Check root readme file

+++ objects/geometry_common.h
@@ -50,10 +50,10 @@ namespace Graphics
 
   struct Color
   {
-    uint8_t a;
-    uint8_t r;
-    uint8_t g;
-    uint8_t b;
+    float a;
+    float r;
+    float g;
+    float b;
   };
 
   struct Triangle_indexed
